custom page builder added

+++ database/migrations/2022_02_25_061634_create_pages_table.php
@@ -20,8 +20,8 @@ class CreatePagesTable extends Migration
             $table->string('slug', 550);
             $table->longText('content')->nullable();
             $table->string('featured_icon')->nullable();
-            $table->decimal('user_id');
-            $table->decimal('status');
+            $table->integer('user_id');
+            $table->integer('status');
             $table->string('lang');
             
             $table->timestamps();
